Conflicts:
	src/Chunk.cpp

+++ src/BlockArea.h
@@ -294,7 +294,7 @@ public:
 	NIBBLETYPE * GetBlockMetas   (void) const { return m_BlockMetas; }     // NOTE: one byte per block!
 	NIBBLETYPE * GetBlockLight   (void) const { return m_BlockLight; }     // NOTE: one byte per block!
 	NIBBLETYPE * GetBlockSkyLight(void) const { return m_BlockSkyLight; }  // NOTE: one byte per block!
-	size_t       GetBlockCount(void) const { return m_Size.x * m_Size.y * m_Size.z; }
+	size_t       GetBlockCount(void) const { return (size_t)(m_Size.x * m_Size.y * m_Size.z); }
 	int MakeIndex(int a_RelX, int a_RelY, int a_RelZ) const;
 
 protected:
